requirements.txt
python-dotenv
jira
ipinfo